Refactor: Use `ignoredProperties` for all languages
