+++ src/language-css/clean.js
@@ -3,19 +3,17 @@
 const { isFrontMatterNode } = require("../common/util");
 const getLast = require("../utils/get-last");
 
-function clean(ast, newObj, parent) {
-  [
-    "raw", // front-matter
-    "raws",
-    "sourceIndex",
-    "source",
-    "before",
-    "after",
-    "trailingComma",
-  ].forEach((name) => {
-    delete newObj[name];
-  });
+const ignoredProperties = new Set([
+  "raw", // front-matter
+  "raws",
+  "sourceIndex",
+  "source",
+  "before",
+  "after",
+  "trailingComma",
+]);
 
+function clean(ast, newObj, parent) {
   if (isFrontMatterNode(ast) && ast.lang === "yaml") {
     delete newObj.value;
   }
@@ -169,6 +167,8 @@ function clean(ast, newObj, parent) {
     delete newObj.value;
   }
 }
+
+clean.ignoredProperties = ignoredProperties;
 
 function cleanCSSStrings(value) {
   return value.replace(/'/g, '"').replace(/\\([^\dA-Fa-f])/g, "$1");
