+++ src/language-handlebars/clean.js
@@ -1,9 +1,6 @@
 "use strict";
 
-module.exports = function (ast, newNode) {
-  delete newNode.loc;
-  delete newNode.selfClosing;
-
+function clean(ast, newNode /*, parent*/) {
   // (Glimmer/HTML) ignore TextNode whitespace
   if (ast.type === "TextNode") {
     const trimmed = ast.chars.trim();
@@ -12,4 +9,7 @@ module.exports = function (ast, newNode) {
     }
     newNode.chars = trimmed;
   }
-};
+}
+clean.ignoredProperties = new Set(["loc", "selfClosing"]);
+
+module.exports = clean;
