+++ src/language-js/printer-estree-json.js
@@ -58,22 +58,27 @@ function genericPrint(path, options, print) {
   }
 }
 
-function clean(node, newNode /*, parent*/) {
-  delete newNode.start;
-  delete newNode.end;
-  delete newNode.extra;
-  delete newNode.loc;
-  delete newNode.comments;
-  delete newNode.errors;
-  delete newNode.range;
+const ignoredProperties = new Set([
+  "start",
+  "end",
+  "extra",
+  "loc",
+  "comments",
+  "errors",
+  "range",
+]);
 
-  if (node.type === "Identifier") {
+function clean(node, newNode /*, parent*/) {
+  const { type } = node;
+  if (type === "Identifier") {
     return { type: "StringLiteral", value: node.name };
   }
-  if (node.type === "UnaryExpression" && node.operator === "+") {
+  if (type === "UnaryExpression" && node.operator === "+") {
     return newNode.argument;
   }
 }
+
+clean.ignoredProperties = ignoredProperties;
 
 module.exports = {
   preprocess,
