+++ src/language-markdown/printer-markdown.js
@@ -25,10 +25,11 @@ const {
   utils: { normalizeDoc },
   printer: { printDocToString },
 } = require("../document");
-const { replaceEndOfLineWith, isFrontMatterNode } = require("../common/util");
+const { replaceEndOfLineWith } = require("../common/util");
 const embed = require("./embed");
-const pragma = require("./pragma");
+const { insertPragma } = require("./pragma");
 const preprocess = require("./preprocess");
+const clean = require("./clean");
 const {
   getFencedCodeBlockValue,
   hasGitDiffFriendlyOrderedList,
@@ -935,74 +936,6 @@ function clamp(value, min, max) {
   return value < min ? min : value > max ? max : value;
 }
 
-function clean(ast, newObj, parent) {
-  delete newObj.position;
-  delete newObj.raw; // front-matter
-
-  // for codeblock
-  if (
-    ast.type === "front-matter" ||
-    ast.type === "code" ||
-    ast.type === "yaml" ||
-    ast.type === "import" ||
-    ast.type === "export" ||
-    ast.type === "jsx"
-  ) {
-    delete newObj.value;
-  }
-
-  if (ast.type === "list") {
-    delete newObj.isAligned;
-  }
-
-  if (ast.type === "list" || ast.type === "listItem") {
-    delete newObj.spread;
-    delete newObj.loose;
-  }
-
-  // texts can be splitted or merged
-  if (ast.type === "text") {
-    return null;
-  }
-
-  if (ast.type === "inlineCode") {
-    newObj.value = ast.value.replace(/[\t\n ]+/g, " ");
-  }
-
-  if (ast.type === "wikiLink") {
-    newObj.value = ast.value.trim().replace(/[\t\n]+/g, " ");
-  }
-
-  if (ast.type === "definition" || ast.type === "linkReference") {
-    newObj.label = ast.label
-      .trim()
-      .replace(/[\t\n ]+/g, " ")
-      .toLowerCase();
-  }
-
-  if (
-    (ast.type === "definition" ||
-      ast.type === "link" ||
-      ast.type === "image") &&
-    ast.title
-  ) {
-    newObj.title = ast.title.replace(/\\(["')])/g, "$1");
-  }
-
-  // for insert pragma
-  if (
-    parent &&
-    parent.type === "root" &&
-    parent.children.length > 0 &&
-    (parent.children[0] === ast ||
-      (isFrontMatterNode(parent.children[0]) && parent.children[1] === ast)) &&
-    ast.type === "html" &&
-    pragma.startWithPragma(ast.value)
-  ) {
-    return null;
-  }
-}
-
 function hasPrettierIgnore(path) {
   const index = +path.getName();
 
@@ -1020,5 +953,5 @@ module.exports = {
   embed,
   massageAstNode: clean,
   hasPrettierIgnore,
-  insertPragma: pragma.insertPragma,
+  insertPragma,
 };
